Fix any base to binary overflow errors and increase max values using Long instead of int [User Stories #142 #143][Tasks #147 #148 #149 #150]

+++ commons/src/main/java/com/simpletools/calculator/commons/operations/BitwiseOperationFactory.kt
@@ -5,7 +5,7 @@ import com.simpletools.calculator.commons.operations.base.* // ktlint-disable no
 
 object BitwiseOperationFactory {
 
-    fun forId(id: String, baseValue: Int, secondValue: Int): BitwiseOperation? {
+    fun forId(id: String, baseValue: Long, secondValue: Long): BitwiseOperation? {
         when (id) {
             NEGATIVE -> return bitwiseUnarySet(BitwiseNegativeOperation, baseValue)
             AND -> return bitwiseBinarySet(AndOperation, baseValue, secondValue)
@@ -16,13 +16,13 @@ object BitwiseOperationFactory {
         }
     }
 
-    private fun bitwiseBinarySet(operation: BinaryBitwiseOperation, baseValue: Int, secondValue: Int): BitwiseOperation? {
+    private fun bitwiseBinarySet(operation: BinaryBitwiseOperation, baseValue: Long, secondValue: Long): BitwiseOperation? {
         operation.baseValue = baseValue
         operation.secondValue = secondValue
         return operation
     }
 
-    private fun bitwiseUnarySet(operation: UnaryBitwiseOperation, value: Int): BitwiseOperation? {
+    private fun bitwiseUnarySet(operation: UnaryBitwiseOperation, value: Long): BitwiseOperation? {
         operation.value = value
         return operation
     }
